+++ unlock-app/src/__tests__/utils/url.test.ts
@@ -1,4 +1,4 @@
-import { rewriteIpfsUrl } from '../../utils/url'
+import { rewriteIpfsUrl, getSlugParamsFromUrl } from '../../utils/url'
 import { it, describe, expect } from 'vitest'
 describe('url', () => {
   it('correctly parse url', () => {
@@ -23,5 +23,39 @@ describe('url', () => {
     const url = 'google.it'
 
     expect(rewriteIpfsUrl(url)).toBe('google.it')
+  })
+})
+
+describe('getSlugParamsFromUrl helper', () => {
+  it('returns hash and params', () => {
+    expect.assertions(4)
+    expect(
+      getSlugParamsFromUrl('/certification#test-event-2?tokenId=2')
+    ).toStrictEqual({
+      hash: 'test-event-2',
+      params: {
+        tokenId: '2',
+      },
+    })
+
+    expect(
+      getSlugParamsFromUrl('/certification#test-event-2?tokenId=2&domain=test')
+    ).toStrictEqual({
+      hash: 'test-event-2',
+      params: {
+        tokenId: '2',
+        domain: 'test',
+      },
+    })
+
+    expect(getSlugParamsFromUrl('/event#test')).toStrictEqual({
+      hash: 'test',
+      params: {},
+    })
+
+    expect(getSlugParamsFromUrl('/event')).toStrictEqual({
+      hash: undefined,
+      params: {},
+    })
   })
 })
